Refactor installed and outdated views with improved loading states and error boundaries
