+++ extensions/brew/src/installed.tsx
@@ -1,11 +1,16 @@
+/**
+ * Installed view for displaying installed brew packages.
+ */
+
 import { useState } from "react";
-import { Cask, Formula } from "./utils/brew";
+import { Cask, Formula, uiLogger } from "./utils";
 import { useBrewInstalled } from "./hooks/useBrewInstalled";
 import { isInstalled } from "./hooks/useBrewSearch";
 import { FormulaList } from "./components/list";
 import { InstallableFilterDropdown, InstallableFilterType, placeholder } from "./components/filter";
+import { ErrorBoundary } from "./components/ErrorBoundary";
 
-export default function Main() {
+function InstalledContent() {
   const [filter, setFilter] = useState(InstallableFilterType.all);
   const { isLoading, data: installed, revalidate } = useBrewInstalled();
 
@@ -18,16 +23,41 @@ export default function Main() {
     casks = Array.from(installed.casks.values());
   }
 
+  // Log rendering statistics
+  if (installed && !isLoading) {
+    uiLogger.log("Installed view rendered", {
+      filter,
+      formulaeDisplayed: formulae.length,
+      casksDisplayed: casks.length,
+      totalDisplayed: formulae.length + casks.length,
+      totalAvailable: (installed.formulae?.size ?? 0) + (installed.casks?.size ?? 0),
+    });
+  }
+
+  // Determine search bar placeholder based on loading state
+  const searchBarPlaceholder = isLoading ? "Loading installed packages…" : placeholder(filter);
+
   return (
     <FormulaList
       formulae={formulae}
       casks={casks}
-      searchBarPlaceholder={placeholder(filter)}
+      searchBarPlaceholder={searchBarPlaceholder}
       searchBarAccessory={<InstallableFilterDropdown onSelect={setFilter} />}
       isLoading={isLoading}
       dataFetched={installed !== undefined}
       isInstalled={(name) => isInstalled(name, installed)}
-      onAction={() => revalidate()}
+      onAction={() => {
+        uiLogger.log("Revalidating installed packages");
+        revalidate();
+      }}
     />
+  );
+}
+
+export default function Main() {
+  return (
+    <ErrorBoundary>
+      <InstalledContent />
+    </ErrorBoundary>
   );
 }
